+++ thermion_dart/lib/src/filament/src/interface/filament_app.dart
@@ -386,7 +386,6 @@ abstract class FilamentApp<T> {
   Future<ThermionAsset> createGeometry(
     Geometry geometry, {
     List<MaterialInstance>? materialInstances,
-    bool releaseSourceData = false,
   });
 
   //